refactor: use only breakpoints file
