--- src/enums/medias.ts
@@ -1,5 +1,0 @@
-export default {
-  minMobile: '360px',
-  maxMobile: '768px',
-  maxTablet: '992px',
-} as const
