+++ src/layout/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { breakpoints } from '../utils/breakpoints'
+import { breakpoints } from '../enums/breakpoints'
 
 export const Container = styled.main`
   color: var(--white);
